Moved functionality from Command spec to Callable
This way the Command module will be cleaner and the functionality will
be in Callable instead.
